+++ io-package.json
@@ -190,7 +190,7 @@
     "updateUnchangedObjects": false,
     "blacklist": [
       {
-        "id": "ABCDEF"
+        "id": "SHBDUO-1#123456789#2"
       }
     ]
   },
