+++ lib/datapoints.js
@@ -223,9 +223,9 @@ function getDeviceByType(type, protocol) {
   let device;
 
   if (getDeviceGen(type) == 2) {
-    device = clone(Object.assign(devices[type], defaultsgen2))
+    device = clone(Object.assign(devices[type], defaultsgen2));
   } else {
-    device = clone(Object.assign(devices[type], defaultsgen1))
+    device = clone(Object.assign(devices[type], defaultsgen1));
   }
 
   device = deleteNoDisplay(device, protocol);
