+++ lib/devices/default.js
@@ -116,12 +116,8 @@ let defaultsgen1 = {
   },
   'hostname': {
     coap: {
-      http_publish: '/status',
-      http_publish_funct: (value) => { return value ? JSON.parse(value).wifi_sta.ip : undefined; }
     },
     mqtt: {
-      mqtt_publish: 'shellies/announce',
-      mqtt_publish_funct: (value) => { return value ? JSON.parse(value).ip : false; }
     },
     common: {
       'name': 'Device Hostname',
@@ -330,20 +326,6 @@ let defaultsgen2 = {
   },
   'hostname': {
     mqtt: {
-      http_publish: '/rpc/Shelly.GetStatus',
-      http_publish_funct: async (value) => {
-        if (value) {
-          const valueObj = JSON.parse(value);
-
-          if (valueObj.wifi.sta_ip) {
-            return valueObj.wifi.sta_ip;
-          } else if (valueObj.eth.ip) {
-            return valueObj.eth.ip;
-          }
-        }
-
-        return undefined;
-      }
     },
     common: {
       'name': 'Device Hostname',
