+++ lib/protocol/base.js
@@ -6,10 +6,48 @@
 'use strict';
 
 const datapoints = require(__dirname + '/../datapoints');
+const request = require('request');
+
+function recursiveSubStringReplace(source, pattern, replacement) {
+    function recursiveReplace(objSource) {
+        switch (typeof objSource) {
+            case 'string':
+                return objSource.replace(pattern, replacement);
+            case 'object':
+                if (objSource === null) {
+                    return null;
+                }
+                Object.keys(objSource).forEach(function (property) {
+                    objSource[property] = recursiveReplace(objSource[property]);
+                });
+                return objSource;
+            default:
+                return objSource;
+        }
+    }
+    return recursiveReplace(source);
+}
+
+function isAsync(funct) {
+    if (funct && funct.constructor) return funct.constructor.name == 'AsyncFunction';
+    return undefined;
+}
+
+function requestAsync(url) {
+    return new Promise((resolve, reject) => {
+        request(url, (error, res, body) => {
+            if (!error && body) {
+                resolve(body);
+            } else {
+                reject(error);
+            }
+        });
+    });
+}
 
 class BaseClient {
     constructor(type, adapter, objectHelper, eventEmitter) {
-        this.type = type;
+        this.type = type;   // mqtt or coap
 
         this.adapter = adapter;
         this.objectHelper = objectHelper;
@@ -22,21 +60,27 @@ class BaseClient {
         this.device = {};
         this.http = {};
 
-        this.devicename;
+        this.devicename;    // e.g. SHBDUO-1#8CAAB5616291#2
         this.ip;            // e.g. 192.168.1.2
         this.auth;          // Authorization HTTP Header
 
-        this.id;
-        this.devicetype;
-        this.deviceid;
-        this.serialid;
-        this.deviceexist;
-        this.devicegen;
+        this.id;            // e.g. ShellyBulbDuo-8CAAB5616291
+        this.devicetype;    // e.g. SHBDUO-1
+        this.deviceid;      // e.g. ShellyBulbDuo
+        this.serialid;      // e.g. 8CAAB5616291
+
+        this.deviceexists;  // boolean
+        this.devicegen;     // 1 or 2
+
         this.httptimeout = 5 * 1000;
 
         if (this.adapter.config.httpusername && this.adapter.config.httppassword && this.adapter.config.httpusername.length > 0 && this.adapter.config.httppassword.length > 0) {
             this.auth = 'Basic ' + Buffer.from(this.adapter.config.httpusername + ':' + this.adapter.config.httppassword).toString('base64');
         }
+
+        // Handle firmware updates
+        this.firmwareUpdatePolling();
+        this.eventEmitter.on('onFirmwareUpdate', async () => await this.firmwareUpdate(true));
     }
 
     /**
@@ -58,27 +102,30 @@ class BaseClient {
     }
 
     /**
-    * Returns a string for Logging with the IP address and name of Shelly Device and type
+    * Get IP of device. For example
+    * 192.168.1.2
+    */
+    getIP() {
+        return this.ip;
+    }
+
+    /**
+    * Returns a string for logging with the IP address and name of Shelly device and type
     */
     getName() {
-        let name = this.getDeviceName(); // SHRGBW2#1234#1
-        let ip = this.getIP(); // 192.168.11.1
-        let deviceid = this.getDeviceId(); // shellyplug-s-12345
-        let id = this.getId(); // shellyplug-s-12345
-        return ip + ' (' + deviceid + ' / ' + id + ' / ' + name + ')';
+        return this.getIP() + ' (' + this.getDeviceId() + ' / ' + this.getId() + ' / ' + this.getDeviceName() + ')';
     }
 
     /**
     * Checks if Shelly device type in the configuration exist. If missing
     * you have to add a configuration in the ./lib/devices direcotory
     */
-    deviceExist() {
-        if (this.deviceexist === undefined) {
-            let deviceid = this.getDeviceId();
-            this.deviceexist = datapoints.getDeviceNameForMQTT(deviceid) ? true : false;
+    deviceExists() {
+        if (this.deviceexists === undefined) {
+            this.deviceexists = datapoints.getDeviceNameForMQTT(this.getDeviceId()) ? true : false;
         }
 
-        return this.deviceexist;
+        return this.deviceexists;
     }
 
     /**
@@ -89,7 +136,6 @@ class BaseClient {
         if (this.devicegen === undefined) {
             let deviceid = this.getDeviceId();
             this.devicegen = datapoints.getDeviceGen(deviceid);
-            this.adapter.log.debug('DEVICE GEN: ' + this.devicegen);
         }
 
         return this.devicegen;
@@ -99,13 +145,8 @@ class BaseClient {
     * Missing data will be pulled by http
     */
     async httpIoBrokerState() {
-        // Test - to delete after a while
-        if (!this.httpIoBrokerStateTime || Date.now() >= (this.httpIoBrokerStateTime + (1000 * 60 * 10))) {
-            this.httpIoBrokerStateTime = Date.now();
-        }
-
-        if (!this.adapter.isOnline(this.getDeviceId())) {
-            this.timerid = setTimeout(async () => await this.httpIoBrokerState(), 100);
+        if (!this.adapter.isOnline(this.getDeviceName())) {
+            this.timerid = setTimeout(async () => await this.httpIoBrokerState(), 1000);
             return;
         }
 
@@ -129,9 +170,10 @@ class BaseClient {
                     };
                 }
 
-                this.adapter.log.debug('http request' + JSON.stringify(params) + ' for ' + this.getName());
+                this.adapter.log.debug(`HTTP request ${JSON.stringify(params)} for ${this.getName()}`);
 
                 let body = await requestAsync(params);
+                // this.adapter.log.debug(`HTTP response ${JSON.stringify(body)} for ${this.getName()}`);
                 for (let j in states) {
                     let state = this.device[states[j]];
                     if (state && state.state) {
@@ -172,7 +214,7 @@ class BaseClient {
                 }
                 this.objectHelper.processObjectQueue(() => { });
             } catch (error) {
-                //
+                this.adapter.log.error(`HTTP request error: ${JSON.stringify(error)}`);
             }
         }
 
@@ -181,14 +223,171 @@ class BaseClient {
         }
     }
 
-    /*
+    /**
+     * Create objects in shelly.X. for a new shelly device
+     * The Shelly device has to exist in the ./lib/devices/ directory
+     */
+    createObjects() {
+        if (Object.keys(this.device).length === 0) {
+            let devicetype = datapoints.getDeviceNameForCoAP(this.getDeviceType());
+            let devices = datapoints.getDeviceByType(devicetype, this.type);
+            if (devices) {
+                devices = recursiveSubStringReplace(devices, new RegExp('<mqttprefix>', 'g'), this.mqttprefix);
+                devices = recursiveSubStringReplace(devices, new RegExp('<devicetype>', 'g'), devicetype);
+                devices = recursiveSubStringReplace(devices, new RegExp('<deviceid>', 'g'), this.getSerialId());
+                for (let j in devices) {
+                    let statename = j;
+                    let state = devices[statename];
+                    state.state = statename;
+                    let deviceid = this.getDeviceName();
+                    if (!this.states[deviceid] || this.states[deviceid] !== deviceid) {
+                        this.states[deviceid] = deviceid;
+                        this.objectHelper.setOrUpdateObject(deviceid, {
+                            type: 'device',
+                            common: {
+                                name: 'Device ' + deviceid
+                            },
+                            native: {}
+                        }, ['name']);
+                    }
+                    let channel = statename.split('.').slice(0, 1).join();
+                    if (channel !== statename) {
+                        let channelid = deviceid + '.' + channel;
+                        if (!this.states[channelid] || this.states[channelid] !== channelid) {
+                            this.states[channelid] = channelid;
+                            this.objectHelper.setOrUpdateObject(channelid, {
+                                type: 'channel',
+                                common: {
+                                    name: 'Channel ' + channel
+                                }
+                            }, ['name']);
+                        }
+                    }
+                    let stateid = deviceid + '.' + statename;
+                    let controlFunction;
+                    if (state[this.type] && state[this.type].http_cmd && !state[this.type].coap_cmd) {
+                        controlFunction = async (value) => {
+                            if (state[this.type] && state[this.type].http_cmd_funct) {
+                                try {
+                                    value = isAsync(state[this.type].http_cmd_funct) ? await state[this.type].http_cmd_funct(value, this) : state[this.type].http_cmd_funct(value, this);
+                                } catch (error) {
+                                    this.adapter.log.error(`Error in function state.${this.type}.http_cmd_funct for state ${stateid} for ${this.getName()}: ${error}`);
+                                }
+                            }
+                            let body;
+                            let params;
+                            try {
+                                if (this.auth) {
+                                    params = {
+                                        url: 'http://' + this.getIP() + state[this.type].http_cmd,
+                                        timeout: this.httptimeout,
+                                        qs: value,
+                                        headers: {
+                                            'Authorization': this.auth
+                                        }
+                                    };
+                                } else {
+                                    params = {
+                                        url: 'http://' + this.getIP() + state[this.type].http_cmd,
+                                        timeout: this.httptimeout,
+                                        qs: value
+                                    };
+                                }
+                                this.adapter.log.debug('Call url ' + JSON.stringify(params) + ' for ' + this.getName());
+                                body = await requestAsync(params);
+                                // this.adapter.log.debug('Create Object body : ' + body);
+                            } catch (error) {
+                                if (body && body === '401 Unauthorized') {
+                                    this.adapter.log.error('Wrong http username or http password! Please enter the user credential from restricted login for ' + this.getName());
+                                } else {
+                                    this.adapter.log.error(`Error in function state.${this.type}.http_cmd for state ${stateid} and request ${JSON.stringify(params)} for ${this.getName()}: ${error}`);
+                                }
+                            }
+                            delete this.states[stateid];
+                        };
+                    }
+                    if (state[this.type] && state[this.type].http_publish && !state[this.type].coap_publish) {
+                        if (!this.http[state[this.type].http_publish]) this.http[state[this.type].http_publish] = [];
+                        this.http[state[this.type].http_publish].push(statename);
+                    }
+                    let value;
+                    if (state[this.type].coap_init_value) value = state[this.type].coap_init_value;
+                    this.objectHelper.setOrUpdateObject(stateid, {
+                        type: 'state',
+                        common: state.common
+                    }, ['name'], value, controlFunction);
+                }
+                this.device = devices;
+            }
+            this.objectHelper.processObjectQueue(() => { });
+        }
+    }
+
+    /**
+     * delete old states in objects unter shelly.X.
+     * For example if the configuration for the device change
+     */
+    async deleteOldStates() {
+        let id = this.adapter.namespace + '.' + this.getDeviceName();
+        let obj = await this.adapter.getAdapterObjectsAsync();
+        let devicetype = datapoints.getDeviceNameForCoAP(this.getDeviceType());
+
+        let dps = datapoints.getAll(this.type);
+        dps = dps[devicetype];
+
+        if (dps) {
+            for (let i in obj) {
+                let tmpid = obj[i];
+                if (tmpid && tmpid._id && tmpid.type) {
+                    let stateid = tmpid._id.replace(id + '.', '');
+                    if (tmpid.type === 'state' && tmpid._id.startsWith(id)) {
+                        if (!dps[stateid]) {
+                            try {
+                                await this.adapter.delObjectAsync(tmpid._id);
+                                delete obj[tmpid._id];
+                                this.adapter.log.info('Delete old state: ' + tmpid._id);
+                            } catch (error) {
+                                this.adapter.log.error('Cound not delete old state: ' + tmpid._id);
+                            }
+                        }
+                    }
+                }
+            }
+        }
+        // delete empty channels
+        for (let i in obj) {
+            let tmpidi = obj[i];
+            if (tmpidi && tmpidi.type && tmpidi._id && tmpidi.type === 'channel') {
+                let found = false;
+                for (let j in obj) {
+                    let tmpidj = obj[j];
+                    if (!tmpidj) {
+                        continue;
+                    }
+                    if (tmpidj && tmpidj.type && tmpidj._id && tmpidj.type === 'state' && tmpidj._id.startsWith(tmpidi._id)) {
+                        found = true;
+                        break;
+                    }
+                }
+                if (found === false) {
+                    try {
+                        this.adapter.log.info('Delete old channel: ' + tmpidi._id);
+                        await this.adapter.delObjectAsync(tmpidi._id);
+                        delete obj[tmpidi._id];
+                    } catch (error) {
+                        this.adapter.log.error('Could not delete old channel: ' + tmpidi._id);
+                    }
+                }
+            }
+        }
+    }
+
     async firmwareUpdatePolling() {
         if (this.adapter.config.autoupdate) {
             await this.firmwareUpdate(true);
             this.autoupdateid = setTimeout(async () => await this.firmwareUpdatePolling(), 60 * 1000);
         }
     }
-    */
 
     async firmwareUpdate(update) {
         if (!update) return;
@@ -240,7 +439,7 @@ class BaseClient {
             this.devicetype = undefined;
             this.deviceid = undefined;
             this.serialid = undefined;
-            this.deviceexist = undefined;
+            this.deviceexists = undefined;
             this.devicegen = undefined;
         }
     }
