+++ lib/protocol/coap.js
@@ -21,38 +21,6 @@ function isAsync(funct) {
   return undefined;
 }
 
-function requestAsync(url) {
-  return new Promise((resolve, reject) => {
-    request(url, (error, res, body) => {
-      if (!error && body) {
-        resolve(body);
-      } else {
-        reject(error);
-      }
-    });
-  });
-}
-
-function recursiveSubStringReplace(source, pattern, replacement) {
-  function recursiveReplace(objSource) {
-    switch (typeof objSource) {
-      case 'string':
-        return objSource.replace(pattern, replacement);
-      case 'object':
-        if (objSource === null) {
-          return null;
-        }
-        Object.keys(objSource).forEach(function (property) {
-          objSource[property] = recursiveReplace(objSource[property]);
-        });
-        return objSource;
-      default:
-        return objSource;
-    }
-  }
-  return recursiveReplace(source);
-}
-
 /**
  * get the CoAP value by key
  * @param {object/string} key - like 112 or '112' or [11,12] or ['11','12']
@@ -98,8 +66,6 @@ class CoAPClient extends BaseClient {
 
     super('coap', adapter, objectHelper, eventEmitter);
 
-    this.getOldDeivceInfo(devicename);
-
     this.shelly = shelly;
     this.devicename = devicename;
     this.ip = ip;
@@ -111,24 +77,15 @@ class CoAPClient extends BaseClient {
   }
 
   /**
-     * Change Device ID from #2 to #1 in fimrware verson >= 1.8
-     * Example SHRGBW2#1234#1 will change to SHRGBW2#1234#2
-     */
-  getOldDeivceInfo(devicename) {
-    // if(devicename) return devicename.split('#').slice(0, 2).join() + '#1';
+   * Change Device ID from #2 to #1 (fimrware verson >= 1.8)
+   * Example SHRGBW2#1234#2 will change to SHRGBW2#1234#1
+   */
+  getOldDeviceInfo(devicename) {
     if (devicename && devicename.substr(-2) === '#2') {
       return devicename.substr(0, devicename.length - 2) + '#1';
     } else {
       return devicename;
     }
-  }
-
-  /**
-   * Get IP of device back. For example
-   * 192.168.1.2
-   */
-  getIP() {
-    return this.ip;
   }
 
   /**
@@ -138,7 +95,9 @@ class CoAPClient extends BaseClient {
     if (!this.id) {
       let devicetype = datapoints.getDeviceNameForCoAP(this.getDeviceType());
       let serialid = this.getSerialId();
-      if (devicetype && serialid) this.id = devicetype + '-' + serialid;
+      if (devicetype && serialid) {
+        this.id = devicetype + '-' + serialid;
+      }
     }
     return this.id;
   }
@@ -150,7 +109,7 @@ class CoAPClient extends BaseClient {
    * Example: SHRGBW2#1234#1
    */
   getDeviceName() {
-    return this.getOldDeivceInfo(this.devicename);
+    return this.getOldDeviceInfo(this.devicename);
   }
 
   /**
@@ -205,162 +164,6 @@ class CoAPClient extends BaseClient {
     if (this.listenerds) this.shelly.removeListener('device-connection-status', this.listenerds);
   }
 
-  /**
-   * delete old states in objects unter shelly.X.
-   * For example if the configuration for the device change
-   */
-  async deleteOldStates() {
-    let id = this.adapter.namespace + '.' + this.getDeviceName();
-    let obj = await this.adapter.getAdapterObjectsAsync();
-    let devicetype = datapoints.getDeviceNameForCoAP(this.getDeviceType());
-    let dps = datapoints.getAll('coap');
-    dps = dps[devicetype];
-    if (dps) {
-      for (let i in obj) {
-        let tmpid = obj[i];
-        if (tmpid && tmpid._id && tmpid.type) {
-          let stateid = tmpid._id.replace(id + '.', '');
-          if (tmpid.type === 'state' && tmpid._id.startsWith(id)) {
-            if (!dps[stateid]) {
-              try {
-                await this.adapter.delObjectAsync(tmpid._id);
-                delete obj[tmpid._id];
-                this.adapter.log.info('Delete old state: ' + tmpid._id);
-              } catch (error) {
-                this.adapter.log.error('Cound not delete old state: ' + tmpid._id);
-              }
-            }
-          }
-        }
-      }
-    }
-    // delete empty channels
-    for (let i in obj) {
-      let tmpidi = obj[i];
-      if (tmpidi && tmpidi.type && tmpidi._id && tmpidi.type === 'channel') {
-        let found = false;
-        for (let j in obj) {
-          let tmpidj = obj[j];
-          if (!tmpidj) {
-            continue;
-          }
-          if (tmpidj && tmpidj.type && tmpidj._id && tmpidj.type === 'state' && tmpidj._id.startsWith(tmpidi._id)) {
-            found = true;
-            break;
-          }
-        }
-        if (found === false) {
-          try {
-            this.adapter.log.info('Delete old channel: ' + tmpidi._id);
-            await this.adapter.delObjectAsync(tmpidi._id);
-            delete obj[tmpidi._id];
-          } catch (error) {
-            this.adapter.log.error('Could not delete old channel: ' + tmpidi._id);
-          }
-        }
-      }
-    }
-  }
-
-  /**
-   * Create objects unter shelly.0 for a new shelly device
-   * The Shelly device has to exist in the ./lib/devices/ directory
-   */
-  createObjects() {
-    if (Object.keys(this.device).length === 0) {
-      let devicetype = datapoints.getDeviceNameForCoAP(this.getDeviceType());
-      let devices = datapoints.getDeviceByType(devicetype, 'coap');
-      if (devices) {
-        devices = recursiveSubStringReplace(devices, new RegExp('<devicetype>', 'g'), devicetype);
-        devices = recursiveSubStringReplace(devices, new RegExp('<deviceid>', 'g'), this.getSerialId());
-        for (let j in devices) {
-          let statename = j;
-          let state = devices[statename];
-          state.state = statename;
-          let deviceid = this.getDeviceName();
-          if (!this.states[deviceid] || this.states[deviceid] !== deviceid) {
-            this.states[deviceid] = deviceid;
-            this.objectHelper.setOrUpdateObject(deviceid, {
-              type: 'device',
-              common: {
-                name: 'Device ' + deviceid
-              },
-              native: {}
-            }, ['name']);
-          }
-          let channel = statename.split('.').slice(0, 1).join();
-          if (channel !== statename) {
-            let channelid = deviceid + '.' + channel;
-            if (!this.states[channelid] || this.states[channelid] !== channelid) {
-              this.states[channelid] = channelid;
-              this.objectHelper.setOrUpdateObject(channelid, {
-                type: 'channel',
-                common: {
-                  name: 'Channel ' + channel
-                }
-              }, ['name']);
-            }
-          }
-          let stateid = deviceid + '.' + statename;
-          let controlFunction;
-          if (state.coap && state.coap.http_cmd && !state.coap.coap_cmd) {
-            controlFunction = async (value) => {
-              if (state.coap && state.coap.http_cmd_funct) {
-                try {
-                  value = isAsync(state.coap.http_cmd_funct) ? await state.coap.http_cmd_funct(value, this) : state.coap.http_cmd_funct(value, this);
-                } catch (error) {
-                  this.adapter.log.error('Error in function state.coap.http_cmd_funct for state ' + stateid + ' for ' + this.getName() + ' (' + error + ')');
-                }
-              }
-              let body;
-              let params;
-              try {
-                if (this.auth) {
-                  params = {
-                    url: 'http://' + this.getIP() + state.coap.http_cmd,
-                    timeout: this.httptimeout,
-                    qs: value,
-                    headers: {
-                      'Authorization': this.auth
-                    }
-                  };
-                } else {
-                  params = {
-                    url: 'http://' + this.getIP() + state.coap.http_cmd,
-                    timeout: this.httptimeout,
-                    qs: value
-                  };
-                }
-                this.adapter.log.debug('Call url ' + JSON.stringify(params) + ' for ' + this.getName());
-                body = await requestAsync(params);
-                // this.adapter.log.debug('Create Object body : ' + body);
-              } catch (error) {
-                if (body && body === '401 Unauthorized') {
-                  this.adapter.log.error('Wrong http username or http password! Please enter the user credential from restricted login for ' + this.getName());
-                } else {
-                  this.adapter.log.error('Error in function state.coap.http_cmd for state ' + stateid + ' and request' + JSON.stringify(params) + ' for ' + this.getName() + ' (' + error + ')');
-                }
-              }
-              delete this.states[stateid];
-            };
-          }
-          if (state.coap && state.coap.http_publish && !state.coap.coap_publish) {
-            if (!this.http[state.coap.http_publish]) this.http[state.coap.http_publish] = [];
-            this.http[state.coap.http_publish].push(statename);
-          }
-          let value;
-          if (state.coap.coap_init_value) value = state.coap.coap_init_value;
-          this.objectHelper.setOrUpdateObject(stateid, {
-            type: 'state',
-            common: state.common
-          }, ['name'], value, controlFunction);
-        }
-        this.device = devices;
-      }
-      this.objectHelper.processObjectQueue(() => { });
-    }
-  }
-
   getDevices() {
     let states = [];
     for (let i in this.device) {
@@ -377,7 +180,6 @@ class CoAPClient extends BaseClient {
     return states;
   }
 
-
   /**
    * State changes from device will be saved in the ioBroker states
    * @param {object} payload - object can be ervery type of value
@@ -391,18 +193,17 @@ class CoAPClient extends BaseClient {
       let stateid = deviceid + '.' + dp.state;
       let value = payload;
       this.adapter.log.debug('Create State : ' + stateid + ', Payload: ' + JSON.stringify(payload) + ' for ' + this.getDeviceName());
-      this.adapter.log.debug('Create State : ' + stateid + ', Payload: ' + payload.toString() + ' for ' + this.getDeviceName());
       try {
         if (dp.coap && dp.coap.coap_publish) {
           value = getCoapValue(dp.coap.coap_publish, value.G);
           if (dp.coap && dp.coap.coap_publish_funct) {
             value = isAsync(dp.coap.coap_publish_funct) ? await dp.coap.coap_publish_funct(value, this) : dp.coap.coap_publish_funct(value, this);
           }
+
           if (dp.common.type === 'boolean' && value === 'false') value = false;
           if (dp.common.type === 'boolean' && value === 'true') value = true;
           if (dp.common.type === 'number' && value !== undefined) value = Number(value);
-          // this.adapter.log.debug('createIoBrokerState(), State : ' + stateid + ', Value: ' + JSON.stringify(value));
-          // if (value !== undefined && (!Object.prototype.hasOwnProperty.call(this.states, stateid) || this.states[stateid] !== value)) {}
+
           if (value !== undefined && (!Object.prototype.hasOwnProperty.call(this.states, stateid) || this.states[stateid] !== value || this.adapter.config.updateUnchangedObjects)) {
             this.adapter.log.debug('State change : ' + stateid + ', Value: ' + JSON.stringify(value) + ' for ' + this.getName());
             this.states[stateid] = value;
@@ -420,27 +221,34 @@ class CoAPClient extends BaseClient {
     this.objectHelper.processObjectQueue(() => { });
   }
 
-  start(payload, description) {
-    if (this.deviceExist()) {
+  async start(payload, description) {
+    if (this.deviceExists()) {
       let polltime = this.getPolltime();
       if (polltime > 0) {
         this.adapter.log.info('Shelly device ' + this.getName() + ' with CoAP connected! Polltime set to ' + polltime + ' sec.');
       } else {
         this.adapter.log.info('Shelly device ' + this.getName() + ' with CoAP connected! No polling');
       }
+
       this.adapter.log.debug('1. Shelly device Info: ' + this.getDeviceName() + ' : ' + JSON.stringify(description));
       this.adapter.log.debug('2. Shelly device Info: ' + this.getDeviceName() + ' : ' + JSON.stringify(payload));
+
       this.deleteOldStates();
       this.createObjects();
+
+      // Fill hostname
+      await this.adapter.setStateAsync(this.getDeviceName() + '.hostname', {val: this.getIP(), ack: true});
+      this.adapter.emit('deviceStatusUpdate', this.getDeviceName(), true);
+
       this.httpIoBrokerState();
+
       if (payload) this.createIoBrokerState(payload);
-      this.eventEmitter.on('onFirmwareUpdate', async () => await this.firmwareUpdate(true));
-      // this.firmwareUpdatePolling();
+
       this.listener();
     } else {
-      this.adapter.log.error('Shelly Device unknown, configuratin for Shelly device ' + this.getName() + ' for CoAP does not exist!');
+      this.adapter.log.error('Shelly Device unknown, configuration for Shelly device ' + this.getName() + ' for CoAP does not exist!');
       this.adapter.log.error('1. Send developer following Info: ' + this.getDeviceName() + ' : ' + JSON.stringify(description));
-      this.adapter.log.error('2 .Send developer following Info: ' + this.getDeviceName() + ' : ' + JSON.stringify(payload));
+      this.adapter.log.error('2. Send developer following Info: ' + this.getDeviceName() + ' : ' + JSON.stringify(payload));
     }
   }
 
@@ -448,22 +256,20 @@ class CoAPClient extends BaseClient {
     this.shelly.on('error', (err) => {
       this.adapter.log.debug('Error handling Shelly data: ' + err);
     });
+
     this.shelly.on('update-device-status', this.listenerus = (devicename, payload) => {
-      if (devicename.startsWith('SHSW-1#058F87')) {
-        let a = 1;
-      }
-      if (this.getOldDeivceInfo(devicename) === this.getDeviceName()) {
+      if (this.getOldDeviceInfo(devicename) === this.getDeviceName()) {
         this.createIoBrokerState(payload);
       }
     });
+
     this.shelly.on('device-connection-status', this.listenerds = (devicename, connected) => {
       this.adapter.log.debug('Connection update received for ' + devicename + ': ' + connected);
-      if (this.getOldDeivceInfo(devicename) === this.getDeviceName()) {
-        // adapter.setState(deviceId + '.online', connected, true);
+      if (this.getOldDeviceInfo(devicename) === this.getDeviceName()) {
+        this.adapter.emit('deviceStatusUpdate', this.getDeviceName(), true);
       }
     });
   }
-
 }
 
 class CoAPServer extends BaseServer {
@@ -508,43 +314,50 @@ class CoAPServer extends BaseServer {
       };
     } else {
       options = {
-        logger: this.adapter.log.debug,
+        logger: this.adapter.log.debug
       };
     }
+
     if (this.adapter.config.coapbind && this.adapter.config.coapbind != '0.0.0.0') {
       options.multicastInterface = this.adapter.config.coapbind;
     }
+
     let shelly = new Shelly(options);
+
     shelly.on('error', (err) => {
       this.adapter.log.debug('Error handling Shelly data: ' + err);
     });
-    shelly.on('update-device-status', (deviceId, status) => {
-      this.adapter.log.debug('Status update received for ' + deviceId + ': ' + JSON.stringify(status));
-      if (deviceId && typeof deviceId === 'string') {
-        shelly.getDeviceDescription(deviceId, (err, deviceId, description, ip) => {
-          if (!err && deviceId && ip) {
+
+    shelly.on('update-device-status', (devicename, status) => {
+      this.adapter.log.debug(`Status update received for ${devicename}: ${JSON.stringify(status)}`);
+
+      if (devicename && typeof devicename === 'string') {
+        shelly.getDeviceDescription(devicename, (err, devicename, description, ip) => {
+          if (!err && devicename && ip) {
             // if ip address change of coap device change
-            if (this.clients[deviceId] && this.clients[deviceId].getIP() !== ip) {
-              this.clients[deviceId].destroy();
-              delete this.clients[deviceId];
+            if (this.clients[devicename] && this.clients[devicename].getIP() !== ip) {
+              this.clients[devicename].destroy();
+              delete this.clients[devicename];
             }
-            if (!this.clients[deviceId]) {
-              if (!this.isBlackListed(deviceId) && !this.isBlackListed(ip)) {
-                this.clients[deviceId] = new CoAPClient(this.adapter, this.objectHelper, this.eventEmitter, shelly, deviceId, ip, status, description);
+            if (!this.clients[devicename]) {
+              if (!this.isBlackListed(devicename) && !this.isBlackListed(ip)) {
+                this.clients[devicename] = new CoAPClient(this.adapter, this.objectHelper, this.eventEmitter, shelly, devicename, ip, status, description);
               }
             }
           }
         });
       } else {
-        this.adapter.log.debug('Device Id is missing ' + deviceId);
+        this.adapter.log.debug('Device Id is missing ' + devicename);
       }
     });
+
     shelly.on('disconnect', () => {
       for (let i in this.clients) {
         this.clients[i].destroy();
         delete this.clients[i];
       }
     });
+
     shelly.listen(() => {
       this.adapter.log.info('Listening for Shelly packets in the network');
     });
