+++ lib/protocol/mqtt.js
@@ -91,7 +91,7 @@ class MQTTClient extends BaseClient {
   }
 
   /**
-   * to get sure, thatan instance will be start more than one, we check for running instances
+   * to get sure, that an instance will be start more than one, we check for running instances
    * if an instance run with same name (shellyswitch-12345), we destroy the old instance
    * @param {object} self - my instance
    */
@@ -297,62 +297,6 @@ class MQTTClient extends BaseClient {
   }
 
   /**
-   * delete old states in objects unter shelly.X.
-   * For example if the configuration for the device change
-   */
-  async deleteOldStates() {
-    let id = this.adapter.namespace + '.' + this.getDeviceName();
-    let obj = await this.adapter.getAdapterObjectsAsync();
-    let dps = datapoints.getAll('mqtt');
-    let deviceid = this.getDeviceId();
-    dps = dps[deviceid];
-    if (dps) {
-      for (let i in obj) {
-        let tmpid = obj[i];
-        if (!tmpid) continue;
-        let stateid = tmpid._id.replace(id + '.', '');
-        if (tmpid.type === 'state' && tmpid._id.startsWith(id)) {
-          if (!dps[stateid]) {
-            try {
-              await this.adapter.delObjectAsync(tmpid._id);
-              delete obj[tmpid._id];
-              this.adapter.log.info('Delete old state: ' + tmpid._id);
-            } catch (error) {
-              this.adapter.log.error('Could not delete old state: ' + tmpid._id);
-            }
-          }
-        }
-      }
-    }
-    // delete empty channels
-    for (let i in obj) {
-      let tmpidi = obj[i];
-      if (tmpidi && tmpidi.type && tmpidi._id && tmpidi.type === 'channel') {
-        let found = false;
-        for (let j in obj) {
-          let tmpidj = obj[j];
-          if (!tmpidj) {
-            continue;
-          }
-          if (tmpidj && tmpidj.type && tmpidj._id && tmpidj.type === 'state' && tmpidj._id.startsWith(tmpidi._id)) {
-            found = true;
-            break;
-          }
-        }
-        if (found === false) {
-          try {
-            await this.adapter.delObjectAsync(tmpidi._id);
-            delete obj[tmpidi._id];
-            this.adapter.log.info('Delete old channel: ' + tmpidi._id);
-          } catch (error) {
-            this.adapter.log.error('Could not delete old channel: ' + tmpidi._id);
-          }
-        }
-      }
-    }
-  }
-
-  /**
    * Create objects unter shelly.0 for a new shelly device
    * The Shelly device has to exist in the ./lib/devices/ directory
    */
@@ -516,8 +460,6 @@ class MQTTClient extends BaseClient {
 
   start() {
     this.client = mqtt(this.stream);
-    this.eventEmitter.on('onFirmwareUpdate', async () => await this.firmwareUpdate(true));
-    // this.firmwareUpdatePolling();
     this.listener();
   }
 
@@ -570,6 +512,7 @@ class MQTTClient extends BaseClient {
   setMqttPrefixByWill(topic) {
     // Gen1: "shellies/huhu-shellybutton1-A4CF12F454A3/online"
     // Gen2: "shellyplus1pm-44179394d4d4/online"
+
     if (this.mqttprefix) {
       return this.mqttprefix;
     } else {
@@ -596,7 +539,7 @@ class MQTTClient extends BaseClient {
     this.client.on('connect', async (packet) => {
       this.packet = packet;
       this.adapter.log.debug('client connected: ' + JSON.stringify(packet));
-      if (this.deviceExist()) {
+      if (this.deviceExists()) {
         if (packet.username === this.adapter.config.mqttusername && packet.password !== undefined && packet.password.toString() === this.adapter.config.mqttpassword) {
           // check for existing instances
           MQTTClient._registerRun(this);
@@ -606,18 +549,22 @@ class MQTTClient extends BaseClient {
           } else {
             this.adapter.log.info('Shelly device ' + this.getName() + ' with MQTT connected! No polling');
           }
+
           // Letzer Wille speichern
           if (packet.will) {
             this.will = packet.will;
           }
+
           if (this.will && this.will.topic) {
             this.setMqttPrefixByWill(this.will.topic);
           } else {
             await this.setMqttPrefixHttp();
           }
+
           this.deleteOldStates();
           this.createObjects();
           this.httpIoBrokerState();
+
           this.client.connack({ returnCode: 0 });
           // this.client.connack({ returnCode: 0, sessionPresent });
         } else {
